+++ internal/file/parsers/copy.go
@@ -1,4 +1,4 @@
-package parser
+package parsers
 
 import (
 	"strings"
