+++ internal/file/parsers/fkey.go
@@ -1,4 +1,4 @@
-package parser
+package parsers
 
 import "regexp"
